small edit

+++ board_utils/add_piece_to_board.c
@@ -1,7 +1,7 @@
 #include "board_utils.h"
 
 
-void add_piece_to_board(unsigned short board[21], unsigned short piece[4][21], int selected_rot) {
+void add_piece_to_board(unsigned int board[21], unsigned int piece[4][21], int selected_rot) {
     for (int i = 0; i < 21; i++) {
         board[i] |= piece[selected_rot][i];
     }
