+++ board_utils/fall.c
@@ -1,6 +1,6 @@
 #include "board_utils.h"
 
-int fall(unsigned short board[21], unsigned short piece[4][21], int selected_rot) {
+int fall(unsigned int board[21], unsigned int piece[4][21], int selected_rot) {
     for (int i = 0; i < 20; i++) {
         if (board[i + 1] & piece[selected_rot][i]) {
             return 0;
