+++ board_utils/init_game_board.c
@@ -2,7 +2,7 @@
 
 #define BOARD_WITH_WALLS 0b0010000000000100
 
-void init_game_board(unsigned short board[21]) {
+void init_game_board(unsigned int board[21]) {
     for (int i = 0; i < 20; i++) {
         board[i] = BOARD_WITH_WALLS;
     }
