+++ board_utils/init_piece_board.c
@@ -3,7 +3,7 @@
 
 #define piece_t enum piece_t
 
-void init_piece_board(unsigned short piece[4][21], piece_t piece_type) {
+void init_piece_board(unsigned int piece[4][21], piece_t piece_type) {
     for (int i = 0; i < 4; i++) {
         for (int j = 0; j < 21; j++) {
             piece[i][j] = 0b0000000000000000;
