+++ makers/make_i.c
@@ -1,6 +1,6 @@
 #include "makers.h"
 
-void make_i(unsigned short piece[4][21]) {
+void make_i(unsigned int piece[4][21]) {
     piece[0][1] = 0b0000001111000000;
     piece[1][0] = 0b0000000010000000;
     piece[1][1] = 0b0000000010000000;
